TileMatrix und Money fix

+++ Assets/Scripts/ApplicationMenu.cs
@@ -12,17 +12,19 @@ public class ApplicationMenu : MonoBehaviour, IDataPersistence
     public double multi_3;
     public double multi_4;
     public double moneyMultiplikator;
-    public double money;
+    public int money;
     public TMP_Dropdown options_1;
     public TMP_Dropdown options_2;
     public TMP_Dropdown options_3;
     public TMP_Dropdown options_4;
     public TMP_Dropdown degree;
-    public void LoadData(GameData data){
+    public void LoadData(GameData data)
+    {
         this.money = data.money;
     }
 
-    public void SaveData(ref GameData data){
+    public void SaveData(ref GameData data)
+    {
         data.money = this.money;
     }
 
@@ -50,113 +52,114 @@ public class ApplicationMenu : MonoBehaviour, IDataPersistence
         });
     }
 
-     void valueChanged_1(TMP_Dropdown option)
+    void valueChanged_1(TMP_Dropdown option)
     {
         //or about this
-         if(option.value == 1 || option.value == 2)
+        if (option.value == 1 || option.value == 2)
         {
             multi_1 = 0.5;
         }
-        
-        else if(option.value == 3)
+
+        else if (option.value == 3)
         {
             multi_1 = 0.3;
         }
 
-        else if(option.value == 4 || option.value == 6)
+        else if (option.value == 4 || option.value == 6)
         {
             multi_1 = 0.1;
         }
-        
+
         else
         {
             multi_1 = 0.2;
         }
     }
-    
-        void valueChanged_2(TMP_Dropdown option)
+
+    void valueChanged_2(TMP_Dropdown option)
     {
-         if(option.value == 1 || option.value == 2)
+        if (option.value == 1 || option.value == 2)
         {
             multi_2 = 0.5;
         }
-        
-        else if(option.value == 3)
+
+        else if (option.value == 3)
         {
             multi_2 = 0.3;
         }
 
-        else if(option.value == 4 || option.value == 6)
+        else if (option.value == 4 || option.value == 6)
         {
             multi_2 = 0.3;
         }
-        
+
         else
         {
             multi_2 = 0.2;
         }
     }
 
-         void valueChanged_3(TMP_Dropdown option)
+    void valueChanged_3(TMP_Dropdown option)
     {
-         if(option.value == 1 || option.value == 2)
+        if (option.value == 1 || option.value == 2)
         {
             multi_3 = 0.5;
         }
-        
-        else if(option.value == 3)
+
+        else if (option.value == 3)
         {
             multi_3 = 0.3;
         }
 
-        else if(option.value == 4 || option.value == 6)
+        else if (option.value == 4 || option.value == 6)
         {
             multi_3 = 0.3;
         }
-        
+
         else
         {
             multi_3 = 0.2;
         }
     }
-    
-        void valueChanged_4(TMP_Dropdown option)
+
+    void valueChanged_4(TMP_Dropdown option)
     {
-        if(option.value == 1 || option.value == 2)
+        if (option.value == 1 || option.value == 2)
         {
             multi_4 = 0.5;
         }
-        
-        else if(option.value == 3)
+
+        else if (option.value == 3)
         {
             multi_4 = 0.3;
         }
 
-        else if(option.value == 4 || option.value == 6)
+        else if (option.value == 4 || option.value == 6)
         {
             multi_4 = 0.3;
         }
-        
+
         else
         {
             multi_4 = 0.2;
         }
     }
 
-    public void Apply ()
+    public void Apply()
     {
-        if(degree.value == 0 || degree.value == 1)
+        if (degree.value == 0 || degree.value == 1)
         {
             SceneManager.LoadScene("Game Over");
         }
 
-        else {
+        else
+        {
             SceneManager.LoadScene("Management");
         }
 
-        
+
         moneyMultiplikator = multi_1 + multi_2 + multi_3 + multi_4;
-        money = moneyMultiplikator * 40000;
+        money = (int)moneyMultiplikator * 40000;
         DataPersistenceManager.instance.SaveGame();
     }
 }
